Move Setting model to dedicated dynamic_settings app
- Create new apps/dynamic_settings app with Setting model
- Move SettingView and related APIs to dynamic_settings app
- Update imports across codebase to use new location
- Add migration to remove Setting from core app
- Update INSTALLED_APPS in settings
- Move tests to tests/unit/dynamic_settings/
- Remove old Setting-related code from core app
