+++ apps/core/models/__init__.py
@@ -1,6 +1,5 @@
 from .organization import Organization
-from .setting import Setting
 from .team import Team
 from .user import User
 
-__all__ = ["User", "Organization", "Team", "Setting"]
+__all__ = ["User", "Organization", "Team"]
