+++ apps/dynamic_settings/__init__.py
@@ -1,0 +1,5 @@
+"""
+Dynamic Settings app for runtime configuration management.
+"""
+
+default_app_config = "apps.dynamic_settings.apps.DynamicSettingsConfig"
