+++ apps/dynamic_settings/tests/__init__.py
@@ -1,0 +1,1 @@
+"""Tests for dynamic_settings app."""
