+++ apps/dynamic_settings/v1/__init__.py
@@ -1,0 +1,3 @@
+"""
+Dynamic Settings API v1.
+"""
