+++ metrics_service/settings/test.py
@@ -47,6 +47,7 @@ DAB_APPS = [
 
 LOCAL_APPS = [
     "apps.core",
+    "apps.dynamic_settings",
     "apps.tasks",
     "apps.dashboard",
     "apps.api",
